% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{abrw}
\alias{abrw}
\title{The Graphical Laplacian Augmented Screening method for the screening measure}
\usage{
abrw(amatrix, utility, alpha, max_it = 100, threshold = 0)
}
\arguments{
\item{amatrix}{is the adjacency matrix}

\item{utility}{the original ranking measures}

\item{alpha}{the the parameter balances the original ranking measure and the network information}

\item{max_it}{maximum iteration loops.}

\item{threshold}{set the measures smaller than threshold as 0. default 0.}
}
\value{
r: the Graphical Laplacian Augmented measure.
}
\description{
The Graphical Laplacian Augmented Screening method for the screening measure
}
