\name{GLASMcensor-package}
\alias{GLASMcensor-package}
\alias{GLASMcensor}
\docType{package}
\title{\packageTitle{GLASMcensor}}
\description{\packageDescription{GLASMcensor}}
\section{Package Content}{\packageIndices{GLASMcensor}}
\author{\packageAuthor{GLASMcensor}}
\section{Maintainer}{\packageMaintainer{GLASMcensor}}
\keyword{package}
