% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Net_FS}
\alias{Net_FS}
\title{The Personalized PageRank Augmented Screening method for the screening measure}
\usage{
Net_FS(A, Orig_RM, d, r, tol = 1e-06, max_it = 100, threshold = 0)
}
\arguments{
\item{A}{is the adjacency matrix}

\item{Orig_RM}{the original ranking measures}

\item{d}{the the parameter balances the original ranking measure and the network information}

\item{r}{the size of the variables set including all active predictors.}

\item{tol}{the stop iteration condition}

\item{max_it}{maximum iteration loops.}

\item{threshold}{set the measures smaller than threshold as 0. default 0.}
}
\value{
Net_RM: the network-based ranking measure.
}
\description{
The Personalized PageRank Augmented Screening method for the screening measure
}
